+++ web/src/components/ui/index.ts
@@ -632,3 +632,4 @@ export * from './chart-line-volume-osc-cross-sig';
 export * from './chart-line-adx-cross-extreme';
 export * from './chart-line-rsi-overbought-cross';
 export * from './chart-line-rsi-oversold-cross';
+export * from './chart-line-stoch-overbought-cross';